net: shell: Add gptp command
Move "gptp" command to a separate file.
